% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NiNos change(incdec) function.R
\name{change_function}
\alias{change_function}
\title{Change in NiNos function (increase/decrease)}
\usage{
change_function(table_to_use)
}
\arguments{
\item{table_to_use}{Table of data used. Will be mutated with 2 extra columns.}
}
\description{
This function allows you to calculate NiNos percentage change and also print whether they increased 
    or decreased. It will create a percentage column and an inccreasing/deccreasing column.
}
\examples{
change_function()
}
\keyword{NiNos}
