% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NiNos change(updown) function.R
\name{change_function_up_down}
\alias{change_function_up_down}
\title{Change in ninos function (up/down)}
\usage{
change_function_up_down(table_to_use)
}
\arguments{
\item{table_to_use}{Table of data used. Will be mutated with 2 extra columns.}
}
\description{
This function allows you to calculate NiNos percentage change and also print whether
    they went up or down. It will create a column for the percentage change and an up/down column.
}
\examples{
change_function_up_down()
}
\keyword{NiNos}
