% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication cycle function.R
\name{Publication_year_filter}
\alias{Publication_year_filter}
\title{Publication Year Filter Function}
\usage{
Publication_year_filter(data_to_use, Publication_cycle)
}
\arguments{
\item{data_to_use}{Data to check}

\item{Publication_cycle}{Date of publication cycle, will default to "YE SEP 17" as of 3/8/18.}
}
\description{
This function allows you to find where the publication cycle is mentioned in code in order to replace it.
    UPDATE: change the publication_cycle value
}
\examples{
Publication_year_filter()
}
\keyword{cycle}
\keyword{publication}
