% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_th_function.R
\name{reg_th_function}
\alias{reg_th_function}
\title{Registrations in thousands function}
\usage{
reg_th_function(data_to_use, variable_to_count)
}
\arguments{
\item{data_to_use}{Data to be used}

\item{variable_to_count}{Variable to be counted}
}
\description{
This function allows you to count up nino's by year and divide by a thousand to obtain the measure of
"Registrations in thousands"
}
\examples{
reg_th_function()
}
\keyword{NiNos}
